# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Social Message API
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /feeds:
    # binds a127 app logic to a route
    x-swagger-router-controller: feed
    get:
      tags:
      - feed
      description: Returns list of feeds to caller
      # used as the method name of the controller
      operationId: list
      parameters:
      - name: idToken
        in: header
        required: true
        type: string
      responses:
        "200":
          description: A list of feeds.
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/Feed"
        # responses may fall through to errors
        "403":
          description: Forbidden due to access error
          schema:
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    post:
      tags:
      - feed
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: idtoken
        in: header
        required: true
        type: string
      - name: facebookaccesstoken
        in: header
        type: string
      - name: feed
        in: body
        required: true
        schema:
          $ref: "#/definitions/Feed"
      description: Create a feed
      # used as the method name of the controller
      operationId: create
      responses:
        "200":
          description: Newly created feed
          schema:
            $ref: "#/definitions/Feed"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /feed/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: feed
    get:
      tags:
      - feed
      description: Returns the feed for the supplied id to the caller
      parameters:
      - name: idToken
        in: header
        required: true
        type: string
      - name: id
        in: path
        description: Id of the feed to be returned
        required: true
        type: string
      # used as the method name of the controller
      operationId: readById
      responses:
        "200":
          description: A list of feeds.
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/Feed"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    put:
      tags:
      - feed
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: facebookaccesstoken
        in: header
        type: string
      - name: idToken
        in: header
        required: true
        type: string
      - name: id
        in: path
        required: true
        type: string
      - name: feed
        in: body
        required: true
        schema:
          $ref: "#/definitions/Feed"
      description: Update an existing feed
      # used as the method name of the controller
      operationId: updateById
      responses:
        "200":
          description: Updated an existing feed
          schema:
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    delete:
      tags:
      - feed
      description: Deletes the feed for the provided id
      parameters:
      - name: idToken
        in: header
        required: true
        type: string
      - name: id
        in: path
        description: Id of the feed to be deleted
        required: true
        type: string
      # used as the method name of the controller
      operationId: deleteById
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /feed/{id}/scrapNewPostsFromSource:
    x-swagger-router-controller: feed
    post:
        tags:
        - feed
        consumes:
        - application/json
        produces:
        - application/json
        parameters:
        - name: id
          in: path
          required: true
          type: string
        - name: idtoken
          in: header
          required: true
          type: string
        - name: contentText
          in: body
          schema:
            $ref: "#/definitions/GeneralResponse"
        description: Create a feed
        # used as the method name of the controller
        operationId: scrapNewPostsFromSource
        responses:
          "200":
            description: Newly created feed
            schema:
              $ref: "#/definitions/GeneralResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/GeneralResponse"
  /feed/{id}/authorizeToScrap:
    x-swagger-router-controller: feed
    post:
        tags:
        - feed
        consumes:
        - application/json
        produces:
        - application/json
        parameters:
        - name: id
          in: path
          required: true
          type: string
        - name: idtoken
          in: header
          required: true
          type: string
        - name: access_token
          in: header
          required: true
          type: string
        description: Create a feed
        # used as the method name of the controller
        operationId: authorizeToScrap
        responses:
          "200":
            description: Newly created feed
            schema:
              $ref: "#/definitions/GeneralResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/GeneralResponse"
  /feed/{id}/socialPosts:
    # binds a127 app logic to a route
    x-swagger-router-controller: socialPost
    get:
      tags:
      - SocialPost
      description: Returns list of feeds to caller
      # used as the method name of the controller
      operationId: getSocialPosts
      parameters:
      - name: id
        in: path
        required: true
        type: string
      - name: idToken
        in: header
        required: true
        type: string
      responses:
        "200":
          description: A list of socialPosts.
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/SocialPost"
        # responses may fall through to errors
        "403":
          description: Forbidden due to access error
          schema:
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    post:
      tags:
      - SocialPost
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      - name: idtoken
        in: header
        required: true
        type: string
      - name: socialPosts
        in: body
        required: true
        schema:
          type: array
          items:
            $ref: "#/definitions/SocialPost"
      description: Create a feed
      # used as the method name of the controller
      operationId: postSocialPost
      responses:
        "200":
          description: Newly created feed
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/SocialPost"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /socialPost/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: socialPost
    get:
      tags:
      - SocialPost
      description: Returns the feed for the supplied id to the caller
      parameters:
      - name: idToken
        in: header
        required: true
        type: string
      - name: id
        in: path
        description: Id of the feed to be returned
        required: true
        type: string
      # used as the method name of the controller
      operationId: getSocialPostById
      responses:
        "200":
          description: A list of SocialPosts.
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/SocialPost"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    put:
      tags:
      - SocialPost
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: idToken
        in: header
        required: true
        type: string
      - name: id
        in: path
        required: true
        type: string
      - name: socialPost
        in: body
        required: true
        schema:
          $ref: "#/definitions/SocialPost"
      description: Update an existing feed
      # used as the method name of the controller
      operationId: putSocialPostById
      responses:
        "200":
          description: Updated an existing feed
          schema:
            $ref: "#/definitions/PutResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    delete:
      tags:
      - SocialPost
      description: Deletes the feed for the provided id
      parameters:
      - name: idToken
        in: header
        required: true
        type: string
      - name: id
        in: path
        description: Id of the feed to be deleted
        required: true
        type: string
      # used as the method name of the controller
      operationId: deleteSocialPostById
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GeneralResponse:
    required:
      - message
    properties:
      message:
        type: string
  PutResponse:
    required:
      - recordsAffected
      - message
    properties:
      recordsAffected:
        type: integer
      message:
        type: string
  Feed:
    type: object
    required:
      - feedName
      - feedType
    properties:
      uid:
        type: string
      feedName:
        type: string
        maxLength: 50
      feedHandle:
        type: string
        maxLength: 50
      feedType:
        type: string
        default: none
        enum:
        - none
        - facebook
        - twitter
      authentication:
        type: object
      feedStatus:
        type: string
        default: scheduled
        enum:
        - scheduled
        - stopped
      hashTag:
        type: string
        maxLength: 50
  SocialPost:
    type: object
    required:
      - uid
      - feedId
      - profileId
    properties:
      _id:
        type: string
      uid:
        type: string
      feedId:
        type: string
      profileId:
        type: string
      profileName:
        type: string
      profileImageUrl:
        type: string
      postText:
        type: string
      postMediaUrl:
        type: string
      postMediaType:
        type: string
        default: none
        enum:
        - none
        - image
        - video
        - audio
      postedOn:
        type: string
        format: 'YYYY-mm-ddT07:01:36+0000'
      rawPost:
        type: object
      tags:
        type: array
        items:
          type: string
      metaData:
        type: object
      approvalStatus:
        type: string
        default: new
        enum:
        - new
        - approved
        - rejected
        - deleted